% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_esem_loadings}
\alias{plot_esem_loadings}
\title{ESEM rotated factor loading matrix heatmap}
\usage{
plot_esem_loadings(result)
}
\arguments{
\item{result}{A high-level SEMANTICA result or output from \code{ACO_with_ESEM()}.}
}
\value{
A \code{ggplot} object.
}
\description{
ESEM rotated factor loading matrix heatmap
}
\examples{
\dontrun{
plot_esem_loadings(result)
}
}
