% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_loading_profiles}
\alias{plot_loading_profiles}
\title{Factor loading profiles with confidence intervals}
\usage{
plot_loading_profiles(result)
}
\arguments{
\item{result}{Output from \code{ACO_with_ESEM()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Factor loading profiles with confidence intervals
}
\examples{
\dontrun{
plot_loading_profiles(result)
}
}
