% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_score_radar}
\alias{plot_score_radar}
\title{ESEM score decomposition radar chart}
\usage{
plot_score_radar(result)
}
\arguments{
\item{result}{Output from \code{ACO_with_ESEM()}.}
}
\value{
A \code{ggplot} object.
}
\description{
ESEM score decomposition radar chart
}
\examples{
\dontrun{
plot_score_radar(result)
}
}
