% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_dfi_gauges}
\alias{plot_dfi_gauges}
\title{Model fit indices vs DFI cutoffs (gauge chart)}
\usage{
plot_dfi_gauges(result)
}
\arguments{
\item{result}{Output from \code{ACO_with_ESEM()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Model fit indices vs DFI cutoffs (gauge chart)
}
\examples{
\dontrun{
plot_dfi_gauges(result)
}
}
