% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_fitness_evolution}
\alias{plot_fitness_evolution}
\title{Fitness evolution during ACO search}
\usage{
plot_fitness_evolution(result)
}
\arguments{
\item{result}{A high-level SEMANTICA result or output from \code{ACO_with_ESEM()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Fitness evolution during ACO search
}
\examples{
\dontrun{
plot_fitness_evolution(result)
}
}
